fixed delayed animations - see desc
fixing the delays was the easy part -> just add a parameter to the animation state called "Speed" and use a multiplier to manipulate the playback...
but what if you wanted to create the speed variable at runtime?

Through my rabbit hole of research I finally did it, something so small and insignificant that could have easily been replaced by the configuration of the editor, I was determined to see it through in code. ;')